ADEN-7754 Refactor ads module
