+++ app/routes/wiki-page.js
@@ -21,8 +21,7 @@ import {
   namespace as mediawikiNamespace,
   isContentNamespace,
 } from '../utils/mediawiki-namespace';
-import Ads from '../modules/ads/module';
-import waitForAdEngine from '../modules/ads';
+import Ads from '../modules/ads';
 import { logError } from '../modules/event-logger';
 import feedsAndPosts from '../modules/feeds-and-posts';
 
@@ -178,11 +177,11 @@ export default Route.extend(
             !fastboot.get('isFastBoot')
             && !transition.queryParams.noexternals
           ) {
-            waitForAdEngine().then(() => {
+            Ads.waitForAdEngine().then((ads) => {
               model.adsContext.user = model.adsContext.user || {};
               model.adsContext.user.isAuthenticated = this.get('currentUser.isAuthenticated');
 
-              Ads.getInstance().init(model.adsContext);
+              ads.init(model.adsContext);
             });
           }
